refactoring Plans.js
